+++ src/main/java/com/adoonge/seedzip/seed/dto/request/SeedRequest.java
@@ -1,4 +1,4 @@
-package com.adoonge.seedzip.seed.dto.reqeust;
+package com.adoonge.seedzip.seed.dto.request;
 
 import java.time.LocalDate;
