+++ src/main/java/com/adoonge/seedzip/seed/repository/SeedRepositoryCustom.java
@@ -3,7 +3,7 @@ package com.adoonge.seedzip.seed.repository;
 import com.adoonge.seedzip.member.domain.Member;
 import com.adoonge.seedzip.seed.domain.Seed;
 import com.adoonge.seedzip.seed.domain.SeedType;
-import com.adoonge.seedzip.seed.dto.reqeust.SeedFilteringRequest;
+import com.adoonge.seedzip.seed.dto.request.SeedFilteringRequest;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
