+++ src/main/java/com/adoonge/seedzip/seed/service/SeedService.java
@@ -10,8 +10,8 @@ import com.adoonge.seedzip.seed.domain.SeedType;
 import com.adoonge.seedzip.seed.domain.mapping.CategorySeed;
 import com.adoonge.seedzip.seed.domain.mapping.SeedTag;
 import com.adoonge.seedzip.seed.dto.SeedDTO;
-import com.adoonge.seedzip.seed.dto.reqeust.SeedFilteringRequest;
-import com.adoonge.seedzip.seed.dto.reqeust.SeedRequest;
+import com.adoonge.seedzip.seed.dto.request.SeedFilteringRequest;
+import com.adoonge.seedzip.seed.dto.request.SeedRequest;
 import com.adoonge.seedzip.seed.dto.response.SeedResponse;
 import com.adoonge.seedzip.seed.repository.CategorySeedRepository;
 import com.adoonge.seedzip.seed.repository.FileRepository;
